+++ crates/providers/Cargo.toml
@@ -14,6 +14,10 @@ workspace = true
 alloy-eips.workspace = true
 alloy-primitives.workspace = true
 alloy-rpc-types-engine.workspace = true
+kona-providers-alloy = "0.2.0"
+
+# scroll
+scroll-alloy-consensus.workspace = true
 
 # misc
 async-trait.workspace = true
