+++ crates/providers/src/lib.rs
@@ -3,3 +3,6 @@
 
 pub use execution_payload::ExecutionPayloadProvider;
 mod execution_payload;
+
+pub use l1::{L1MessageProvider, L1Provider, OnlineL1Provider};
+mod l1;
